+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
 ### Added
-- `ContainerGetPathCapableTrait`, `ContainerSetPathCapableTrait`.
+- `ContainerGetPathCapableTrait`.
+- `ContainerSetPathCapableTrait`.
+- `ContainerListHasCapableTrait`.
+- `ContainerListGetCapableTrait`.
 
 ## [0.1-alpha6] - 2018-03-03
 ### Added